ensure-sphix: upgrade setuptools in venv
With [1] pip was ensured to be latest, but for
cases like ubuntu bionic which has too old
setuptools(39.0.1) and that doesn't support version
from file, the support only available with
setuptools-39.2.0[2].

This patch proposes to upgrade setuptools
along with pip to support dependencies which
set's version in a file. Other option is to
pass option --use-deprecated legacy-resolver
to pip install or downgrade pip to working version.

This will resolve issue with translation jobs
in neutron which fails during installation of
pyroute2[3].

[1] https://review.opendev.org/c/zuul/zuul-jobs/+/828441
[2] https://github.com/pypa/setuptools/pull/1359/commits/a960ee1c3
[3] https://zuul.openstack.org/builds?job_name=propose-translation-update&project=openstack%2Fneutron

Change-Id: I23475a48fa9cc924962f2f4b5edc7aeb5dd6261c