Donot Setup RDO gpg keys and repo for RHEL and Fedora
Currently in multi-node-bridge role, For RHEL and CentOS
distro, RDO repos are setted up and from where rdo-openvswitch
get pulled in RHEL deployment and causes unwanted failures.

Like a particular version of rdo-openvswitch is not yet available
in CentOS and it fails the job with nothing provides message.

Enabling RDO repos specifically for CentOS and other distros
except RHEL and Fedora will fix the
issue.

Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: Id68f5904c4ec3d667a16e9f4f195d53b02d29cec