Correct git config item name in mirror-workspace-git-repos
It appears this setting inadvertently had an extra word "ignore"
included in the key name, and later versions of the git module in
Ansible (5) throw an error in that case.  Remove it so that this
role will work in all versions.

This has been previously tested in a base-test cycle via
test-mirror-workspace-git-repos.

Change-Id: I597aa8d331e53955e826f09aa51a579e12c94f85