Revert "Install venv for all platforms in ensure-pip"
This reverts commit 0aca37281458b94b30fd124f885cadc18ad3c61e.

This uses the variable ensure_pip_from_upstream_interpreters to
install python-venv packages.

ensure_pip_from_upstream_interpreters is documented as being a list of
python interpreters to run "get-pip.py" with to install pip globally.
This only happens if ensure_pip_from_upstream is set to True.

It's unclear from the original discussion in the change at least, what
this was trying to fix.  It's also unclear why this would skip
packages, and what circumstances that is required in.  It's now
causing more confusion as we try to add fixes ontop of that.

Revert this -- whatever this is doing, we should find the right way to
do it, which isn't this.

Change-Id: Ieada7ec2d2c7320d7aacb9a5dc6de52394c11ed1