ensure-java role: Bump default java_version to 11
Premier Support for Java 8 was discontinued on 31 March 2022: [1]. Even
though Extended Support for Java 8 is still available until 2030, the
vast majority of projects switched to Java 11 already and forced to pass
java_version: 11.

Bump the default and ask the projects that still rely on Java 8 to pass
java_version: 8 to downgrade and use outdated Java releases.

[1] https://www.oracle.com/java/technologies/java-se-support-roadmap.html

Change-Id: I57d08039937d89881c7331293568dc510cabad19