emit-job-header: noqa on error ignore
This marks this particular ignore_errors as safe.  Generally I think
the rule is good, in that tasks get marked as failed when using
ignore_errors which is confusing and hard to debug.  But in this case
it's because of missing hostnames
(I8d67d8aa284b6ce9ae012608e8f1b12784ce836b) and not something we can
catch easily.

Change-Id: I9b5d6a85254f689e97bca91951632b4dea60cfde