test-requirements: bump to Ansible 2.8
We are outside the 4 week period that we dropped Ansible 2.7
(I13802db3314450ad149fdadacd1e2e70dd8468ef) so update the testing
version to 2.8 here.

The comment has become stale, as ansible-lint is now installed via
linters-requirements.txt; remove the bit about how it will pull in the
latest version of Ansible.  Also clarify slightly the version usage.

Later versions of Ansible only have paramiko support as a plugin.  We
have just one place we use it explicitly to generate some keys for
testing; so it is really a dependency of zuul-jobs.  Add it
explicitly.

Change-Id: Iad8622594a773af78487dec88e4ce5cc48806ded