Use RDO wallaby repo to install openvswitch in CS9
We should use the oldest release where we run centos stream 9 jobs which
is wallaby. We set up xena in the past as we backported cs9 jobs to xena
first but now we should switch to wallaby or we may hit issues in
wallaby jobs when the ovs version in xena is higher that in wallaby (as
now).

Change-Id: I18dab9701e96b25396b45df221f54fe94c1ab36b