zuul_azure_storage_upload: rename
roles/upload-logs-azure/tasks/main.yaml calls
"zuul_azure_storage_upload:" but the library file is currently called
"zuul_azure_upload.py".  Since it says in the comments of the file to
call it as "zuul_azure_storage_upload.py" (and that matches the google
one) rename it.

I found this when working backwards with an ansible-lint that runs
against Ansible 2.8.  I think this is an ansible-lint bug; see [1]

[1] https://github.com/ansible/ansible-lint/issues/2283

Change-Id: Ic30d82771e6c591cf17bcd15ca9dc92fb0f89e04