configure-mirrors: make each compontent in 9-stream configurable
When we added 9-stream with I2da3a5e8a45598c6b3ded132ea61b86b4480e262
we had some components pointing to the mirrors and some not.  This is
basically because the OpenDev mirrors choose to mirror these parts of
the distribution.  It was an oversight to do things like this, as this
is nominally a generic configuration tool.

This proposes that we put each component behind a variable which can
be enabled or disabled.  This way users of the role can switch
on-or-off whatever their local mirrors carry, and fall-back to the
generic mirror metalink in other cases.

I imagine that with future distros like this we would have everything
disabled by default in zuul-jobs, and leave it up to sites to
explicitly opt-in the various components.  We could announce this via
our usual deprecation path and do similar for 9-stream; but for now
this keeps the defaults as it is currently written.

Change-Id: Ibfd88c137855c22b960099cc8d582b241c3b1db6