Test speedup change to prepare-workspace-git
The child change to prepare-workspace-git explains this in more detail.
We're reducing the number of ansible loops these roles run to reduce
ansible task startup overhead and speed up the role. This change is made
first to the test role so that we can ensure this all works properly
before affecting production zuul installs.

Change-Id: I3d33bb0001e65f449e1026ed323514097088f664