Update to ansible-lint 6.8.2
This updates to ansible-lint 6.8.2.

The prior changes have updated various things for new things the
linter picked up.

As noted inline, the jinja2 parsing has not been working well with
this repository.  I've been keeping an eye on it over several
releases, but I think at this point it's not going to work well for
us.  I've left discussion in .ansible-lint.

The var-spacing in stage-output is the old name for the jinja2
filtering; since that is now ignored in .ansible-lint it can be
removed.

Change-Id: Ia2c9392eeb232b9b2b1d4febce8493d71de64482