Make test-registry-buildset-registery-k8s-<docker|crio> non-voting
These tests are unhappy after we switched the default nodeset to
Ubuntu Jammy; it is not obvious to me why.  Rather than just pinning
them back to the old distro and leaving it to bitrot more, make these
non-voting as we investigate.

zuul-jobs-test-registry-buildset-registry is also failing, but has to
run for the Zuul to +1 this.  Pin this to focal, but as mentioned
inline we have a follow-on update to skopeo install that seems to fix
this.

Change-Id: Id6ee7ed38fec254493a2abbfa076b9671c907c83